[IMP] website: allow multiple values for visibility conditions
Steps to Reproduce:
1. Go to Website → Edit Mode.
2. Drag and drop a "Contact Us" form.
3. Add two new fields:
    - A selection field (e.g., "Author").
    - A text field.
4. Attempt to set the visibility of the text field such that it only
appears when the user selects multiple values. (e.g., "Mitchell Admin"
and "Marc Demo").

Users cannot select multiple values for visibility conditions when the
comparator is `Contains/Doesn't Contain`, and its behavior is similar to
`Is Equal To/Is Not Equal To`.

This commit introduces support for selecting multiple values when the
visibility comparator is set to `Contains/Doesn't Contain`, provided
the dependency field type is `select`, `checkbox` or `radio`.

task-4203938