[FIX] bus, mail: display a warning when the bus connection is unstable
Gevent server issues can be difficult to detect as there is no visible
hint in the UI to inform the user when something is wrong. As a
result, users may take longer to notice issues with their
applications, leading to support tickets that lack the precision
needed for rapid issue identification.

This PR introduces the `bus_monitoring_service`, a utility designed to
detect unstable or lost bus connections. It displays a small warning
in the UI to promptly alert users of potential connection problems.

The bus connection is considered:
- `LOST` if no connection could be established for more than 5 seconds.
following a reconnect attempt.
- `UNSTABLE` if the two last connections were shorter than 2 minutes.
- `STABLE` once a connection reaches the `STABLE_CONNECTION_TRESHOLD`.

task-4403259