[FIX] sale_project: fix analytic distribution auto set on accounts payable line
Steps to reproduce:
- Install sale, project, and accountant.
- Allow Analytic Accounting from settings.
- Open project application.
- Create a project with a billable feature.
- Add invoices in the topbar.
- Select invoices and create a new invoice.
- Add a product to the invoice (you will see analytic is added).
- Go to the journal items tab.

Issue:
Analytic distribution is already added to the account payable line.
This issue is also produced from the vendor bill and expense.

Cause:
This is due to the functionality to set analytic distribution when an
invoice is created from a project topbar.

Solution:
Do not add the analytic distribution when the account move line is payable.

task-4377544