[IMP] website_sale: review cart_lines images
Depending on the images, sometimes the background color of the image was
matching the background of the checkout page making it hard to discern
the image. To match condensed layout, the images now have a smaller size
on smaller viewports.

Clicking on the image was sometimes redirecting to the product
(suggested products) while regular products weren't. Now all images are
redirecting to the proper page.

task-4387239

images