[FIX] web_editor: remove double unnecessary slashes in tel protocol
Since [1], the tel:// format was preferred over tel: due to reasons
stated in [2]. However, the change in [2] was a workaround for a bug in
the Python library used to parse URLs. According to the relevant RFC,
the tel: protocol should not include double slashes.

When clicking on a link like `tel://45122` in a browser, a popup may
appear attempting to handle the communication using `http://tel://45122`
as the browser does not recognize the tel protocol when double slashes
are included.

This commit removes the unnecessary double slashes to comply with RFC
recommendations and use urllib to parse urls.

[1]: https://github.com/odoo/odoo/commit/6d4a3b3ab5c0f3361d1d681d05b974e295dcbabe
[2]: https://github.com/odoo/odoo/commit/56ce29e71f75a657d4b518d86c1d7084891e898e

task-4331070