[IMP] website: allows to customize the size of s_dynamic_snippet entries
Prior to this commit, the minimum size of entries was based on a
calculation. This works with the default card template but would not
work with other layout types (eg. a list-group).

This commit allows to customize the size of entries for other layouts
defined for `s_dynamic_snippet`.

task-4317830