[FIX] hr_homeworking: fixes to work location display on calendar
- In calendar, when only one employee (different from the user) was
selected as an attendeed in the calendar sidebar, the working
location of this employee was not displayed. To work around this,
switch to `multiCalendar` in that case.

- When the current user was not selected as an attendee, it was possible
to add a working location for the user on a day even if there was
already an exisiting one. To avoid this, always fetch the user's
working location to check if one already exists for a given day, but
only display them if the user's filter is active.

- Work locations were not displayed when "Everybody's calendar" was
selected because the elements of `attendeeIds` had the wrong type.

opw-4373391