[FIX] web: work around timing issue in desktop calendar test
I don't understand that one: in chrome 131 new headless, this test
fails because the dropdown menu never opens.

But it works in the old headless mode as well as when watching the
test run un a browser, the test succeeds.

I assume this is a race in the hooking up of event handlers on the
widget but...