[FIX] point_of_sale: correct reflection of modeupdate
Steps to reproduce:
- Install pos_restaurant.
- Open a table and order something.
- Reopen the table.

Issue:
Even after ordering the order button will be enabled.

Cause:
The presets are not enabled & the default preset_id is set to 0
to notice any changes in preset which default result into modeupdate.

Fix:
The default value is updated based on the selected preset_id &
checking if presets are enabled than only reflect the modeupdates.