[IMP] purchase,sale,stock: add partner warning boxes to form views
Instead of popups, partner warnings/instructions should appear in the
respective form view inside an alert box. This is a warning box in
sale and purchase, containing the partner and the products that
carry warnings, and an info box in stock, containing the partner and
its parent company if either of them has any instructions set.

In addition, the stock picking operation report has a section
displaying the picking instructions, if they were set.

Finally, in sale and purchase order views, the order lines are marked
with the warning colour if the products have warnings.

Task ID: 4268689