[FIX] web: fixed edit_tags behaviour on edit and change in FormView Dialog
In `edit_tags`before when we we make changes in the record, the logic is done as
The changes in record are saved in Form View and in edit_tags the record is
forget and saved again using `X2ManyCrud`, but in here when saving a X2Many
change it is checked in cache and brought back, while cache contains the initial
version of the record, so all the further X2Many changes are invalidated as
X2Many values are retrieved from the cache.

In form view dialog when change the company and try save the record and then
save view we get a traceback saying pushStateBeforeReload is not a function, but
as FormViewDialog not action_window, we dont need to store the state of view, so
we skip saving the state here.

task-3976577