[IMP] website_sale: fix Spacing Issue with contact us and price tags
Steps to Reproduce:
Enable the "Prevent Sale of 0 Product" setting in eCommerce.
Create a product with:
->Zero price
->Price tags.

Issue :
When the "Prevent Sale of 0 Product" setting was enabled, there was a lack of
proper spacing between the "Contact Us" button and the product price tags,
resulting in a misaligned layout.

Fix:
Added a margin-top style to the "price tags" container to ensure
proper spacing between the price tags and contact us button.

Affected Versions:
17.0

opw-4422462