[IMP] mrp: assigned work order to employee in back orders
Before this Commit:
---------------------------------------
When creating backorders for a Manufacturing Order (MO) with Work Orders,
the assigned employees were not carried over. Users had to manually reassign
employees to each backorder, which was time-consuming and error-prone.

After this Commit:
---------------------------------------
This issue occurred due to missing information during the creation of
backorders. The process has now been improved to ensure that the assigned
employees are correctly propagated to backorders, saving time and reducing
the chances of errors.

Task-id: 4405697