[IMP] project: exclude archived users when duplicating or recurring tasks
Prior to this commit, duplicating or recurring a task results in archived user
being assigned to the new task and its associated sub tasks.
This change ensures that archived users are excluded from the assignee of the
new task and sub tasks when a task is duplicated or repeated.

task-4419770