[IMP] event: send email asynchronously by default
Sending emails synchronously has the drawback of causing major
unavailability issues in case of popular events. As a workaround, an
asynchronous parameter was introduced in #155777.

Having the emails sent synchronously by default remains a problem: the
unavailability starts as soon as the users register massively, and it is
impossible to access to the backend in order to change the
configuration.

We change the default behavior to be asynchronous, and we also make sure
to trigger the Email Queue Manager. This way, the delay is reduced to
the minimum.