[IMP] sale: display unpublished combo's item images
Similarly to other types of products, combo products must be published to make them available in eCommerce. However, combo item products (i.e. the different choices in a combo) don't need to be published. Unfortunately, if a product is unpublished, portal users don't have access to its images, so they won't be shown in the combo configurator.

This is a bad UX, as portal users are expected to make a product selection, but we don't show them the corresponding images.

This commits uses the `src` attribute of the `img` tag to show the raw combo item's image instead of using the image url, bypassing access checks.

This is done only for combo items that are not published.

task-4337641