[REF] mail: fetch_mail to use cron progress
Use the `_notify_progress` API for cron execution. At every execution,
read a maximum number of messages and let the cron executor relaunch the
check in a next loop. This also solves a timeout issue when the IMAP
server contains too many new messages.
At every loop, we connect and for each fetched message: retrieve it,
process it in a savepoint, commit and delete or mark as read.

We can refactor the common parts between the server checking logic.
Since we commit after each message, we should ensure the consistency by
adding a savepoint when treating messages.