[IMP] base: cron direct trigger uses _process_job
Behave as a worker and prevent the cron job from messing up the current
HTTP transaction. To do so, process the job the same way as a worker
would do it; cron_cr is the current cursor and let the the processing
create and manage its own cursor for the execution of the job.

task-4433571