[FIX] stock{,_barcode}: trigger quality check on picking validation
Issues:

Creating and validating a picking from the barcode app does not behave
correctly with respect to quality check trigger and creation. Depending
on your flow, you might end up with QC created twice or with QC not
being triggered before validation.

Cause of the issues:

The quality checks are created during the "_action_confirm" of moves:
https://github.com/odoo/enterprise/blob/dd4a1ec4aaa7cbb95d6e40a8bb6ab1d6c88fc7b7/quality_control/models/stock_move.py#L12-L15
However, depending on where you click in the barcode app, the moves
might end up being assigned and hence skip the "_action_confirm" of the
picking (which is computed to be assigned) prior to the
`_pre_action_done_hook` of the `button_validate` of the picking wich is
suppose to return already created quality check:
https://github.com/odoo/odoo/blob/08f0c6481a190349369ead5ab9328886392ddbeb/addons/stock/models/stock_picking.py#L1145-L1147
In which case, the quality check is not triggered (Issue 1).
Furthermore, since  the `_action_confirm` of moves can also be applied
to extra moves created to update the demand and to be merged to the
initial move during the validation, the quality check creation might be
called twice on a move once during its own `action_confirm` and once
during the `_action_confirm` of its extra move (Issue2).

Issue 1:

1. Create a storable product P with a barcode: XXX
2. Go to Quality > Quality Control > Quality Points > New
3. Create a new quality point for your product:
    - Control per: "Operation"
    - Operations: "Receipts"
4. Go to the barcode app > Operations > Receipts > New
5. Scan your product
6. Click on the pencil (key step)
7. Either edit the quantity from the digipad and confirm or go back to
  the picking and edit it from there.
8. Validate the receipt

> Even thought a quality check was created using your quality point for
> your receipt, it was not triggered before validation.

Cause of the Issue:

Clicking on the pencil will trigger a call of the `save_barcode_data`
which will in turn create a stock move line related to the picking
during the write performed on the picking:
https://github.com/odoo/enterprise/blob/dd4a1ec4aaa7cbb95d6e40a8bb6ab1d6c88fc7b7/stock_barcode/controllers/stock_barcode.py#L68-L74
After this edition of the quantity of the picking will write on the
`qty_done` dummy field of that move line which will in turn trigger
a call of its inverse method and modify its quantity:
https://github.com/odoo/enterprise/blob/dd4a1ec4aaa7cbb95d6e40a8bb6ab1d6c88fc7b7/stock_barcode/models/stock_move_line.py#L58-L61
Changing the quantity of the ml will then update the state of the move
to "assigned" because of these lines:
https://github.com/odoo/odoo/blob/a74684777c27e26639b75b0668bb15afadb78c4f/addons/stock/models/stock_move_line.py#L487-L488
https://github.com/odoo/odoo/blob/a74684777c27e26639b75b0668bb15afadb78c4f/addons/stock/models/stock_move_line.py#L530-L531
https://github.com/odoo/odoo/blob/08f0c6481a190349369ead5ab9328886392ddbeb/addons/stock/models/stock_move.py#L2022-L2023
In turns, the state compute method of the picking will determine that
the picking is assigned and the picking will not be confirmed during
its `button_validate` prior to the `_pre_action_done_hook` of the
`button_validate` as it is not in draft anymore.

Fix:

The state of the move should stay as draft just as if we didn't created
the move line through the pencil if we were to edit the qty_done via
the barcode app.

Note:

The inverse method of the "qty_done" computed field was introduced in
17.0+ by commit: 4f19668c51a90e8c12e57b6353bbe5194f83ca10 during the
quantity refactoring so that the issue is not reproducible prior to that
version.

Issue 2:

Same steps 1 -> 5.
6'. click on the +1 button.
7'. Validate > this opens the quality check
8'. Pass the quality check > Validate again
9'. Go to Quality > Quality Control > Quality check

> Two quality checks were created isntead of one for your receipt

Cause of the issue:

At some later point in the button validate of the picking, an
`_action_done` is called to put the pickings and its move in done state:
https://github.com/odoo/odoo/blob/a74684777c27e26639b75b0668bb15afadb78c4f/addons/stock/models/stock_picking.py#L1157

However, during the `_action_done` of the moves if the quantity of the
move exceeds its `product_uom_qty`, extra moves will be created,
confirmed and merged back to the move in order to update its
`product_uom_qty`:
https://github.com/odoo/odoo/blob/607d63fd4d9c07bac9cb68b2d31d86551e93d5b1/addons/stock/models/stock_move.py#L1885-L1890
https://github.com/odoo/odoo/blob/607d63fd4d9c07bac9cb68b2d31d86551e93d5b1/addons/stock/models/stock_move.py#L1823-L1832
However, the action confirm of this extra move will trigger the same
override of the `_action_confirm` that creates a quality check and it
will even be called on the original move that was already confirmed and
already created its QC since the extra move was merged into it:
https://github.com/odoo/enterprise/blob/dd4a1ec4aaa7cbb95d6e40a8bb6ab1d6c88fc7b7/quality_control/models/stock_move.py#L12-L15
As a result we end up with an extra quality check creation.

Fix:

We avoid quality check creation during extra move confirmation.

opw-4266053