[ADD] l10n_in_hr_work_entry_holidays: create work entries for sandwich leaves
In this commit,
we addressed several issues related to the management of leave durations:

- A bug was fixed that incorrectly counted the duration of leave taken
on non-working days, such as Saturdays and Sundays, which previously
resulted in a duration of -2 days. This has now been resolved.

- When creating a leave request that overlaps with non-working days
or public holidays, this commit ensures that non-working days are
included in the leave calculation.

There are several scenarios to consider:

1) When creating leave that covers non-working days, a work entry will be
generated for those days, starting from the next working day.
The duration will be based on the average duration of the working schedule.

2) If leave is created that falls within a public holiday period,
the work entry type for those public holidays will be updated to reflect
the leave's work entry type.

3) If you create individual leaves around non-working days or public holidays,
this will also fall under sandwich leave rules.
Accordingly, work entry will be generated.

Task-4430044