[IMP] {crm, im}_livechat: simplify live chat consultation from lead
This PR aims to improve live chat consultations for leads created by
a chatbot. Specifically, it enhances the formatting of the lead
descriptions to make the conversation easier to read, introduces a
smart button to redirect users to the discussion channel where the
lead originated and set the lead name according to the user request.

task-4432728