[IMP] account: memo generation for grouped payments
--Purpose--

When we create Group Payments (as in selecting multiple invoices, use Pay, and selecting "Group Payments"), it creates one Payment for multiple invoices. In this case, a new Payment memo is generated to easily identify the Payment back. We use "BATCH/XX" as a prefix, but it's a bit confusing as it's not a Batch Payment

--Implementation--

Use "GROUP/XX" in this case instead.

task-4356414