[FIX] hr_contract: fix payslip without hr_contract
The new year 2025 is showing that the contract from user admin will
dynamically end on the current month, making the payslip generation of
june fail when executed in the beginning of the year.

Looks like this was introduced with #61912 checking that the payslip is
only done on a valid contracty.

Multiple possible fix exist but having a demo data with a dynamic end
date while start date is static looks like a good source of date error,
changing the contract to be fixed for the complete year.
(could still break when both modules are not installed at the same time,
on two different years)