[IMP] account: detect duplicate bank accounts
If your database already has a bank account linked to a specific partner,
it will usually be incorrect that another, completely unrelated partner
shares the exact same account.

In some cases it is normal, and we have to allow it (factoring cases,
or specific partners with the same debt collection office like governments),
but it's more the exception than the norm. A warning is a good way to show
the user it exists, but he can ignore it.

task-id 4431041