[REF] portal,l10n*: merge `my/account` and address logic
The purpose of this commit is to properly strong parsing and validation to
ensure smooth address creation and editing from the portal. The address
validation currently used on the My Account page is too old and does not
check all necessary elements.

This commit removes the `my/account` logic and related code and instead
utilizes the address logic to provide a similar user experience and
validation across all address validations in the portal. This approach
will enhance the user experience and make the system more robust.

This commit also improves the related code in e-commerce localization by
fetching specific data only where needed, rather than fetching it for
all "My" pages. Additionally, it refactors the POS invoice generation logic
to reuse the functionality implemented in the portal, reducing the need
for workarounds.

task-3628329