[IMP] {purchase_,sale_,}stock: prevent fractional qtys with serials
Trying to sell, purchase or transfer products tracked by serial
numbers in fractional amounts is nonsensical. It leads to unexpected
effects like the quantity of invoiced products being silently rounded
down to the nearest integer, and breakage during serial number
generation.

This commit adds warnings to sale and purchase views to prevent the
entry of fractional values. It also adds rounding logic for stock
reservations to make sure we always reserve a sufficient amount when a
fractional demand comes in. Tests are also adjusted accordingly for
this new behaviour.