[FIX] base_import: fix selector with an empty string or 0 value
Steps to reproduce:
- Install Accounting
- On Accounting dashboard, click on "Import File" button of
Bank journal
- Select a CSV file with bank statements
- On the left menu, select "No Separator" as "Thousands Separator"

Issue:
"Comma" is displayed as selected instead the first time.
The second time "No Separator" stays as selected.
However, the value sent when testing or importing is not the
correct one.

Cause:
"No Separator" should have an empty string as value.
However, during the generation of the "select" element, the
value for the "option" element is evaluated with:
"opt.value or opt"
As "opt.value" is the empty string (evaluated to False), "opt" is
used instead, even if it is an object, which is not correct.
The main issue is that the empty string or 0 are not handled as
a valid value.

Solution:
Handle the empty string and 0 as acceptable values.

opw-4325310