[FIX] web: export when subfields have properties field
**Steps to reproduce**
1. Have Stock and Sales installed.
2. Go to the Sales list view of products, have more than 1 page of products.
3. Select all the products on the current page by checking the checkbox
on the first line. After that, click on "Select all" to select all
products.
4. Actions > Export
5. Select "Products" (`product_variant_ids`) and then "Products/Stock Quant"
(`stock_quant_ids`).

ValueError: Invalid field stock.quant.sale_ok in leaf ('sale_ok', '=', True)

**Cause**
The domain is invalid for the model. This happens when a model has properties
fields and is not directly the model exported.

**Change**
The domain provided to `/export/get_fields` represents the
domain of the records we are exporting, it should only apply
in that case and not when making a following call to
`getExportedFields` to get the export fields of a related model.

opw-4408457