[FIX] *: minor fixes
- Incorrect "Sent" Status:
When generating the XML with the UBL checkbox selected but without checking the
Nilvera option, the move was incorrectly marked as "Sent" (is_move_sent).  This
behavior has been corrected.

- Traceback During Send and Print with Nilvera:
In the flow where the XML is first generated with UBL, and then the Send and
Print action is performed with Nilvera, Nilvera needed the issues date of the
document to accept it.

- City Name and Subdivision Inversion in XML
The values in the CitySubDivisionName and CityName elements of the XML were
inverted. This has been corrected to align with the expected structure.

task-4457092