[IMP] auth_signup: do not spam the user on new connection
Purpose
=======
When a user has TOTP enabled, on its first login, he receives 2 emails
- "New Connection" email, sent when he enters his credentials
- "Trusted Device Added", if he checked "Don't ask again on this device"

The purpose of that task is to merge both emails when possible

Technical
=========
Because the 2 emails are sent at different moment, we need to delay the sending of
the "New Connection" email.

They are 3 different scenarios:
- chooses to remember the device, the "New Connection" email is already sent (> 1 minute)
  => we send an email saying that a device has been added (so 2 emails are still sent)
- chooses to remember the device, the "New Connection" email is not yet sent (< 1 minute)
  => we merge both emails and force the sending
- chooses to not remember the device
  => we force the sending of the "New Connection" email

Note that if a user enters his credentials, but do not enter the TOTP code,
the sending might take more than 1 minute, because we need the CRON to run
(shouldn't be a big issue, since he didn't log in).

Task-4070794