[IMP] website, *: add Extra info option to searchbar results
*website_sale
Before this commit, there was no option to display product reviews in
the search results. This commit introduces a new generic option for
displaying extra information in search results. For products, this will
show the reviews (if available), but the option can be used for other
entities as well.
For example:
- Blogs -> Display tags
- Events -> Display location or sold-out status
- Appointments -> Display duration time
...

After this commit, product reviews will be shown in the search results
when the option is enabled (if reviews are available).

Part of task-4029264.