[IMP] website: add template option to searchbar snippet
Before this commit, there was no option to customize the template for
how search results are displayed in the search bar snippet.

After this commit, a new option allows users to change the result
template.
Currently, only one additional template is available for demonstration
purposes, but more templates are expected to be added in the future.

Part of task-4029264.