[FIX] point_of_sale: use transition acccount for caba taxes
The reconciliation of the PoS move lines creates a cash basis (caba) move anyway.
Since the real account and not the transition account is used for the caba taxes
the amount from the cash basis move and the PoS move are not matched correctly.
Thus an exchange difference move for the tax amount is created.

opw-4355124