[IMP] l10n_my_edi: Embedded QR code and self billing
As required by the Malaysian government, this
commit will adapt the myinvois sending flows
to allow generating and displaying the QR code
pointing to the portal on the invoice PDF.

This is done in a way that the previous flow
still works until the module is -u and the views
are updated, and then the new flow with the QR code
will be used instead.

It is using a slightly unethical way of storing the
id required for the flow, by using the existing
l10n_my_error_document_hash field which is not used
in case the invoice is valid (our use case here).
This is done to avoid requiring to do an api call
when generating the QR code.
While the field isn't meant for that, this shouldn't
cause any issue and can be changed in master.

--

Self-billing is another document type supported by the MyInvois System.

It is used when the Supplier (e.g., Foreign Suppliers) could not issue
the invoice on MyInvois System. Hence, Buyer issues invoice instead of the supplier.

This commit adapt the flows slightly in order to enable this behavior.

Also removes the "Reject" button from the bill
view, as it's intended to only be used when
importing bills which is a flow that we don't
yet support.

task-4363294