[IMP] l10n_id_efaktur: DPP calculation changes
As of January 1, the Indonesian government has officially increased the tax rate to 12%.
However, due to protests, they decided to revert the tax back to effectively 11%. Since the law has been passed, we can't simply change tax rate to 11%.
Alternatively, the government wants the tax base calculation to be amount applied by tax (DPP) to be multiplied by 11/12.
As a result, in Odoo, the tax configuration will have invoice repartition ilne to contain 91.67% (11/12) of tax instead of the usual 100%.
Due to do this, we have to alter the DPP calculation specifically on eFaktur.

4457091