[FIX] web_editor: fix sidebar overflow issue Step to reproduce:
- set the odoo backend language to spanish (tested in V16.0)
- go to edit mode
- bug: the sidebar menu overflows, there is a horizontal scrollbar.

Issue:

There was an issue with the dynamic width of the buttons within the
button group.The buttons text was overflowing and causing the sidebar
to overflow.

Solution:

Now we have improved css and now the if the button text is too long
it will be truncated and will show ellipsis.

task-4378522