[FIX] mail: chat/channel action menu position
Purpose of this Commit:
This commit resolves the issue of the action menu popover and the notification
settings popover overflowing off the screen. Previously, before this PR, the
popovers could flip to all four directions, but now they are restricted to only
two directions. However, in some cases, the correct position is not among
these two directions. This commit addresses and fixes this behavior.

task-4454161