[IMP] website, *: knowledge article emoji instead of fa book in search
* website_knowledge
Before this commit, when a knowledge article appeared in the search
results, if an image or icon was displayed for that result, it was
represented with the fa_book icon.

Now, when an article is shown, it is represented with the emoji defined
specifically for that article.

Part of task-4029264.